[Docker Configuration] Add configuration for simple deployment
These two files make it easy for end users to deploy the jukebox.
The new compose file knows where the images are and takes arguments as
environment variables. These can be set in compose, which is a wrapper
for docker-compose that injects these variables and specifies the new
compose file.